Configured cloudlet publisher and subscriber for the proper QoS. Implemented a basic configuration with custom QoS data (needs work!)

+++ cloudsim-plus-examples/src-gen/ddsgen/Cloudlet.java
@@ -19,7 +19,7 @@ public class Cloudlet   implements Copyable, Serializable{
 
     public int ttl = (int)0;
     public int peNumber = (int)0;
-    public int utilizationModelParam = (int)0;
+    public double utilizationModelParam = (double)0;
     public String modelType= (String)""; /* maximum length = (255) */
 
     public Cloudlet() {
